feat!(cli): add `-k`/`--stringify-keys` flag
BREAKING CHANGE: make it no longer the default to stringify keys when
converting to JSON. This is being done to avoid silent incorrect
conversions and to turn them into errors. For example, GitHub Actions
workflows with an `on` YAML key are currently converted incorrectly.

Support stringifying keys for TOML.

v0.17.0