perf(toml): make it extremely faster using rtoml
As you can see in https://github.com/pwwang/toml-bench, the slowest toml implementation in Python is tomlkit (the one we were using before), and the fastest one is rtoml.

This change aims to replace tomlkit for rtoml.

The main drawback is that tomlkit is the only style-preserving implementation available (it tries to keep diff minimal). However, that doesn't seem like a big need for remarshal.

@moduon MT-1075