refactor!: put all format options in classes
This removes the formatting parameters to `remarshal.remarshal`
and makes it clear which format uses which options.

With these changes:

- Customizing the output requires one extra function call.
- Conversion with the default formatting
  is a single function call like before.

I am not very concerned about breaking the Python API.
Remarshal users use it through the CLI.
I haven't heard from API users despite one major API break in the past
and have found none on GitHub.
(There is one project that vendors an old version of Remarshal
with some functionality removed, so it won't be affected.)